+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/blocs/pin_code_bloc.dart
@@ -20,17 +20,18 @@ abstract class PinCodeBlocEvents {
 
   ///Triggered when user tap on biometrics button
   void biometricsButtonPressed();
-
-  void checkIfPinIsStored();
 }
 
 /// A contract class containing all states of the PinCodeBloc.
 abstract class PinCodeBlocStates {
-  ///Indicating how much dots to present
+  /// Emits current PIN length to UI
   Stream<int> get digitsCount;
 
+  /// Emits placeholder PIN length to UI
+  Stream<int> get placeholderDigitsCount;
+
   ///Indicating the biometrics button must be presented
-  ConnectableStream<bool> get showBiometricsButton;
+  Stream<bool> get showBiometricsButton;
 
   ///Emits when user successfully authenticate (pin/biometrics)
   ConnectableStream<dynamic> get authenticated;
@@ -48,40 +49,39 @@ class PinCodeBloc extends $PinCodeBloc {
     required this.biometricAuthenticationService,
     required this.pinCodeService,
     required this.localizedReason,
+    required this.autoPromptBiometric,
   }) {
     authenticated.connect().addTo(_compositeSubscription);
-    showBiometricsButton.connect().addTo(_compositeSubscription);
-
-    _$biometricsButtonPressedEvent
-        .switchMap((_) => _authenticateWithBiometrics().asResultStream())
-        .setResultStateHandler(this)
-        .whereSuccess()
-        .listen((_) {})
-        .addTo(_compositeSubscription);
+    Rx.merge([
+      _$addDigitEvent
+          .asyncMap((digit) async => (
+                digit: digit,
+                allowedLength: await pinCodeService.getPinLength()
+              ))
+          .where((data) => _pinCode.value.length <= data.allowedLength)
+          .map((data) => _pinCode.value + data.digit),
+      _$deleteDigitEvent
+          .map((_) => _pinCode.value.substring(0, _pinCode.value.length - 1)),
+    ]).listen(_pinCode.add).addTo(_compositeSubscription);
   }
 
   final PinBiometricsService biometricAuthenticationService;
   final PinCodeService pinCodeService;
   final String localizedReason;
+  final bool autoPromptBiometric;
 
   final BehaviorSubject<String> _pinCode = BehaviorSubject.seeded('');
-  final BehaviorSubject<dynamic> _authResponse = BehaviorSubject();
 
   @override
-  Stream<int> _mapToDigitsCountState() => Rx.merge([
-        _$addDigitEvent.switchMap((digit) => _addDigit(digit).asResultStream()),
-        _$deleteDigitEvent.switchMap(
-          (_) {
-            _pinCode.add(_pinCode.value.substring(
-                0, _pinCode.value.isNotEmpty ? _pinCode.value.length - 1 : 0));
-            return Stream.value(_pinCode.value.length);
-          },
-        ).asResultStream(),
-        errorState
-            .delay(const Duration(milliseconds: 1000))
-            .mapTo(0)
-            .asResultStream(),
-      ]).whereSuccess().startWith(0).share();
+  Stream<int> _mapToDigitsCountState() =>
+      _pinCode.map<int>((pinCode) => pinCode.length).share();
+
+  @override
+  Stream<int> _mapToPlaceholderDigitsCountState() => pinCodeService
+      .getPinLength()
+      .asResultStream()
+      .setResultStateHandler(this)
+      .whereSuccess();
 
   @override
   Stream<bool> _mapToIsLoadingState() => loadingState;
@@ -90,21 +90,33 @@ class PinCodeBloc extends $PinCodeBloc {
   Stream<ErrorModel> _mapToErrorsState() => errorState.mapToErrorModel();
 
   @override
-  ConnectableStream<dynamic> _mapToAuthenticatedState() =>
-      _authResponse.publish();
-
-  @override
-  ConnectableStream<bool> _mapToShowBiometricsButtonState() => Rx.merge([
-        _$checkIfPinIsStoredEvent.switchMap(
-            (_) => pinCodeService.isPinCodeInSecureStorage().asResultStream()),
-        _digitsCountState.switchMap((digitsCount) =>
-            _checkPin(_pinCode.value, digitsCount).asResultStream()),
-        _getAreBiometricsEnabled().asResultStream(),
+  ConnectableStream<dynamic> _mapToAuthenticatedState() => Rx.merge([
+        digitsCount
+            .switchMap((_) => pinCodeService.getPinLength().asStream())
+            .where(
+                (storedPinLength) => storedPinLength == _pinCode.value.length)
+            .switchMap((digitsCount) => _checkPin(_pinCode.value)
+                .asResultStream()
+                .doOnData((result) => _pinCode.add(''))),
+        _$biometricsButtonPressedEvent
+            .mapTo(true)
+            .startWith(autoPromptBiometric)
+            .where((shouldPrompt) => shouldPrompt)
+            .asyncMap((_) => _getAreBiometricsEnabled())
+            .where((biometricsEnabled) => biometricsEnabled)
+            .switchMap((_) => _authenticateWithBiometrics().asResultStream())
       ]).setResultStateHandler(this).whereSuccess().publish();
 
   @override
+  Stream<bool> _mapToShowBiometricsButtonState() => const Stream.empty()
+      .startWith(null)
+      .switchMap((_) => _getAreBiometricsEnabled().asResultStream())
+      .setResultStateHandler(this)
+      .whereSuccess()
+      .shareReplay(maxSize: 1);
+
+  @override
   void dispose() {
-    _authResponse.close();
     _pinCode.close();
     super.dispose();
   }
@@ -130,51 +142,25 @@ class PinCodeBloc extends $PinCodeBloc {
         pinCode != null;
   }
 
-  /// Adds a digit to the pin code, returning the new length of the pin code
-  Future<int> _addDigit(String digit) async {
-    final pinLength = await pinCodeService.getPinLength();
-    if (_pinCode.value.length < pinLength) {
-      _pinCode.add(_pinCode.value + digit);
-    }
-    return _pinCode.value.length;
-  }
-
-  /// Encrypts and verifies the provided pin code
-  Future<dynamic> _encryptAndVerify(String pinCode) async {
+  // Checks the validity of the pin code
+  Future<dynamic> _checkPin(String pinCode) async {
     final encryptedPin = await pinCodeService.encryptPinCode(pinCode);
-    final verifiedPin = await pinCodeService.verifyPinCode(encryptedPin);
-    return verifiedPin;
-  }
-
-  /// Checks the validity of the pin code
-  Future<bool> _checkPin(String pinCode, int digits) async {
-    final storedPinLength = await pinCodeService.getPinLength();
-    if (digits == storedPinLength) {
-      try {
-        final authValue = await _encryptAndVerify(pinCode);
-        final isSaved = await pinCodeService.isPinCodeInSecureStorage();
-        if (isSaved) {
-          _authResponse.add(authValue);
-          return true;
-        }
-      } catch (_) {
-        _pinCode.value = '';
-        rethrow;
-      }
-    }
-    return false;
+    await pinCodeService.savePinCodeInSecureStorage(encryptedPin);
+    final verificationResult = await pinCodeService.verifyPinCode(encryptedPin);
+    return verificationResult;
   }
 
   /// Authenticates the user with biometrics after which the pin code is
   /// retrieved from the device and checked.
-  Future<bool> _authenticateWithBiometrics() async {
+  Future<dynamic> _authenticateWithBiometrics() async {
     if (!await biometricAuthenticationService.isDeviceSupported) {
       throw ErrorEnableBiometrics(BiometricsMessage.notSupported);
     }
     if (await biometricAuthenticationService.authenticate(localizedReason)) {
       final pinCode = await pinCodeService.getPinCode();
       if (pinCode != null) {
-        return await _checkPin(pinCode, pinCode.length);
+        final verificationResult = await pinCodeService.verifyPinCode(pinCode);
+        return verificationResult;
       }
     }
     return false;