+++ packages/widget_toolkit_pin/lib/src/lib_pin_code_with_biometrics/di/pin_code_dependencies.dart
@@ -5,7 +5,6 @@ import 'package:widget_toolkit_biometrics/widget_toolkit_biometrics.dart';
 
 import '../../../widget_toolkit_pin.dart';
 import '../data_source/biometrics_disabled_local_data_source.dart';
-import '../data_source/pin_biometrics_auth_data_source.dart';
 import '../repositories/pin_biometrics_repository.dart';
 import '../services/pin_biometrics_service.dart';
 
@@ -14,22 +13,30 @@ class PinCodeDependencies {
     this.pinCodeService,
     this.biometricsLocalDataSource,
     this.localizedReason,
+    this.biometricsAuthDataSource,
+    this.autoPromptBiometric,
   );
 
   factory PinCodeDependencies.from({
     required PinCodeService pinCodeService,
     required BiometricsLocalDataSource? biometricsLocalDataSource,
     required String localizedReason,
+    PinBiometricsAuthDataSource? biometricsAuthDataSource,
+    bool autoPromptBiometric = false,
   }) =>
       PinCodeDependencies._(
         pinCodeService,
         biometricsLocalDataSource,
         localizedReason,
+        biometricsAuthDataSource,
+        autoPromptBiometric,
       );
 
   final String localizedReason;
   final PinCodeService pinCodeService;
   final BiometricsLocalDataSource? biometricsLocalDataSource;
+  final PinBiometricsAuthDataSource? biometricsAuthDataSource;
+  final bool autoPromptBiometric;
 
   late List<SingleChildWidget> providers = [
     ..._localAuthentication,
@@ -47,9 +54,11 @@ class PinCodeDependencies {
 
   late final List<SingleChildWidget> _dataSources = [
     Provider<PinBiometricsAuthDataSource>(
-      create: (context) => PinBiometricsAuthDataSource(
-        localAuthentication: context.read<LocalAuthentication>(),
-      ),
+      create: (context) =>
+          biometricsAuthDataSource ??
+          PinBiometricsAuthDataSource(
+            localAuthentication: context.read<LocalAuthentication>(),
+          ),
     ),
     if (biometricsLocalDataSource == null)
       Provider<BiometricsLocalDataSource>(
@@ -83,6 +92,7 @@ class PinCodeDependencies {
         biometricAuthenticationService: context.read<PinBiometricsService>(),
         pinCodeService: pinCodeService,
         localizedReason: localizedReason,
+        autoPromptBiometric: autoPromptBiometric,
       ),
     ),
   ];